Prompt for installation directory when running `setup.ps1`
This makes it possible to use the setup script to install multiple instances of RAWeb. This is also useful for people who want to easily customize the RAWeb path. It does not support installing to the root directory of the IIS website. It removes invalid characters from the installation directory. It allows specifying a folder subpath (e.g. apps/raweb).