fix: do not use default icon for apps with dots in their name
The previous logic would detect the presence of a dot in the file name and assumed that indicated the start of the image file extension. For example, for Paint.NET, RAWeb would request https://server/raweb/get-image.aspx?image=resources/Paint.NET&format=png, and the code would assume that the file extnesion was .NET. However, the file is actually named Paint.NET.png.
The new logic checks if the path exists first, and then tries to infer the file extension if the path does not exist. With the Paint.NET example, it would check if the file resources/Paint.NET exists, determine it does not, and then check if resources/Paint.NET.png or resources/Paint.NET.ico exists.