fix: do not specify `FileAssociationIcons`  if the icon does not exist
Before, it always included an IconRaw element with a URL to the ICO file for the file extension, even if the icon for the file extension did not exist. IconRaw should not exist when the file extension icon does not exist. Always including this icon meant that the file extensions received the default icon even when the RemoteApp may have its own icon.