fix: ignore duplicate resources in the same folder
Resources that repeat multiple times the same folder (e.g., file1.rdp and file2.rdp with same file contents) should be ignored after the first resource is included. For schema version 1, this means using the folder path instead of the file path when generating the GUID. For schema version 2, this means checking whether the folder has already been included in the list of folders.