add support for registry images to `get-iamge.aspx.cs`
If the relative file name is prefixed with "registry:", the file name will be treated as the name of the remoteapp in the registry. If it is prefixed with "registry:ext:", the icon for the a file type association (located in the registry) will be used instead of the application's icon.