chore: move app settings from `Web.config` to `App_Data\appSettings.config`
This change means that we can only grant access to app data rather than all of the app's files, including Web.config. Long term, it is better to remove write access to non-user-provided files so that a vulnerability in the raweb code does not allow code that executes on the server to be modified into something malicious.