feat: place security error info bar above all page content
This ensures that the bar does not re-animate on every page change. Additionally, this makes it possible to preoprly vertically center content on a page minus the info bar (when space is available). For example, the notice about no favorites is centered on the page without a scrollbar after this change.