feat: add stale-while-revalidate user cache behavior, allowing the the user cache to be used while the user data is updated in the background if `UserCache.Enabled` is set to `"true"`
With this change, instead of retrieving the user information from the principal before responding, the the user cache can be used for quicker responses. This is important in cases where a domain controller is slow to respond because it is at a remote site.

By default, the cache will only be used for a user if the user data has been updated within the last 59 seconds. It can be modified by setting `UserCache.StaleWhileRevaldate` to a positive integer. It can be disabled by setting `UserCache.StaleWhileRevaldate` to 0 or a negative integer.

The user cache is updated in the background for a particular user once the user has not made a request for at least 5 seconds. Logic is contained within `AuthUserStaleWhileRevalidate.cs`.