chore: add debug and task configurations
New tasks for connecting to IIS Express make it possible to debug. The debugger connects to IIS Express, and IIS Express uses a compiled version of RAWebServer that comes with debug symbols.

THere is also a configuration added for debugging in chrome. It launches a separate chrome profile with debugging mode enabled, which allows breakpoints in vscode to break execution for the vue app. It also allows for interacting with logs from the node portion of vite.