feat: use `TSWorkspace/2.0` for the `User-Agent` when fetching the workspace and its resources
Windows RADC and Windows App both use the `TSWorkspace/2.0` `User-Agent` header when connecting to the workspace and download RDP files and icons. With this PR, the web app replicate that behavior.

`User-Agent` is no longer a forbidden header, so we are allowed to set this header in the browser. Chromium currently does not suppot setting the header, but they are close to implementing this change (10 years later!). FIrefox and Safari already support custom `User-Agent` headers.