feat: support high-resolution ICO files
High resolution ICO files (up to 256x256) contain a mixture of bitmaps at lower resolutions and PNG at higher resolutions. RAWeb was able to read the bitmaps and PNGs, but it was only able to server bitmap-only ICO files. With this change, RAWeb will now correctly read and serve the PNG part of ICO files.